// packages/core/src/schema.generated/tvdbApiSchema.ts
export interface paths {
  "/login": {
    parameters: {
      query?: never
      header?: never
      path?: never
      cookie?: never
    }
    get?: never
    put?: never
    /** create an auth token. The token has one month validation length. */
    post: {
      parameters: {
        query?: never
        header?: never
        path?: never
        cookie?: never
      }
      requestBody: {
        content: {
          "application/json": {
            apikey: string
            pin?: string
          }
        }
      }
      responses: {
        /** @description response */
        200: {
          headers: {
            [name: string]: unknown
          }
          content: {
            "application/json": {
              data?: {
                token?: string
              }
              status?: string
            }
          }
        }
        /** @description invalid credentials */
        401: {
          headers: {
            [name: string]: unknown
          }
          content?: never
        }
      }
    }
    delete?: never
    options?: never
    head?: never
    patch?: never
    trace?: never
  }
// ...
}